# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: YAFA VANAM Commerce API
  version: 2.0.0
  description: |
    The full HTTP contract of the Go commerce API: catalogue, carts, orders,
    Razorpay payments and session auth.

    Conventions:

    - Commerce resources live under `/api/v1`; session auth lives at the root
      (`/auth/...`) because the storefront proxies it separately.
    - Errors use two shapes. Commerce handlers return
      `{"error":{"code","message"}}`; auth handlers return `{"error":"message"}`.
    - Browser mutations while signed in must send `X-CSRF-Token` (double-submit
      with the `yafa_csrf` cookie obtained from `GET /auth/csrf`).
    - Money is a decimal number in INR (e.g. 2119.0); Razorpay amounts are paise.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/products/{slug}:
    get:
      tags: [Catalogue]
      summary: Get an active product with its variants
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: yafa_access
      description: HTTP-only JWT session cookie issued by /auth/login, /auth/register, /auth/refresh, or the Google callback.
    orderToken:
      type: apiKey
      in: header
      name: X-Order-Access-Token
      description: Guest order credential returned as access_token when the order is created (no-auth mode only).
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - not_found
                - access_denied
                - validation_error
                - variant_unavailable
                - invalid_request
                - invalid_idempotency_key
                - payment_order_not_found
                - payment_unavailable
                - payment_provider_error
                - payment_setup_error
                - payment_verification_error
                - invalid_payment_signature
                - invalid_webhook_signature
                - webhook_unavailable
                - service_unavailable
                - unauthenticated
                - unauthorized
                - internal_error
            message: { type: string }
...
    Description:
      type: object
      required: [short, full]
      properties:
        short: { type: string }
        full: { type: string }
    CommerceMoney:
      type: object
      required: [currency, base_price, compare_at_price]
      properties:
        currency: { type: string }
        base_price: { type: number, minimum: 0 }
        compare_at_price: { type: [number, "null"], minimum: 0 }
    Shade:
      type: object
      required: [name, code, hex, undertone]
      properties:
        name: { type: string }
        code: { type: [string, "null"] }
        hex: { type: [string, "null"] }
        undertone: { type: [string, "null"] }
    Variant:
      type: object
      required: [id, sku, size, shade, price, stock, is_active]
      properties:
        id: { type: string }
        sku: { type: [string, "null"] }
        size: { type: [string, "null"] }
        shade: { anyOf: [{ $ref: "#/components/schemas/Shade" }, { type: "null" }] }
        price: { type: number, minimum: 0 }
        stock: { type: [integer, "null"], minimum: 0 }
        is_active: { type: boolean }
    Images:
      type: object
      required: [primary, gallery, lifestyle, detail, texture, alt, paths_verified]
      properties:
        primary: { type: [string, "null"] }
        gallery: { type: array, items: { type: string } }
        lifestyle: { type: array, items: { type: string } }
        detail: { type: array, items: { type: string } }
        texture: { type: [string, "null"] }
        alt: { type: string }
        paths_verified: { type: boolean }
    Usage:
      type: object
      required: [how_to_use, amount, when]
      properties:
        how_to_use: { type: string }
        amount: { type: [string, "null"] }
        when: { type: array, items: { type: string } }
    Ingredients:
      type: object
      required: [full_inci, active_ingredients, ingredient_data_note]
      properties:
        full_inci: { type: [string, "null"] }
        active_ingredients: { type: array, items: {} }
        ingredient_data_note: { type: [string, "null"] }
    Product:
      type: object
      required:
        [id, name, slug, brand, category, subcategory, product_type, status, description, commerce, variants, images, benefits, usage, warnings, ingredients]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        brand: { type: string }
        category: { type: string }
        subcategory: { type: string }
        product_type: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        description: { $ref: "#/components/schemas/Description" }
        commerce: { $ref: "#/components/schemas/CommerceMoney" }
        variants:
          type: array
          items: { $ref: "#/components/schemas/Variant" }
        images: { $ref: "#/components/schemas/Images" }
        benefits: { type: array, items: { type: string } }
        usage: { $ref: "#/components/schemas/Usage" }
        warnings: { type: array, items: { type: string } }
        ingredients: { $ref: "#/components/schemas/Ingredients" }
...